FasterRCNN/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
# Note: detectron2 must be installed separately (see README.md)
# pip install 'git+https://github.com/facebookresearch/detectron2.git'
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.19.0
tqdm>=4.60.0
pyyaml>=5.4.0

